requirements.txt
pandas
numpy
matplotlib
streamlit
asn1crypto==1.5.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==36.0.2
filelock==3.8.0
idna==3.3
oscrypto==1.3.0
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pytz==2022.2.1
requests==2.28.1
typing_extensions==4.3.0
urllib3==1.26.12
snowflake-connector-python==2.7.12
-r https://raw.githubusercontent.com/snowflakedb/snowflake-connector-python/v2.7.12/tested_requirements/requirements_310.reqs
python-dotenv